# internal/api/swagger.yaml
openapi: 3.0.3
info:
  title: Search Trends API
  description: API сервиса для получения Top-N популярных поисковых запросов за последние 5 минут.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /stop-words:
    get:
      tags:
        - StopWords
      summary: Получить список стоп-слов
      operationId: getStopWords
      responses:
        '200':
          description: Список стоп-слов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StopWordsResponse'

    post:
      tags:
        - StopWords
      summary: Добавить слово в стоп-лист
      operationId: addStopWord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddStopWordRequest'
      responses:
        '201':
          description: Слово добавлено в стоп-лист
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StopWordResponse'
        '400':
          description: Некорректное тело запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Слово уже существует в стоп-листе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    AddStopWordRequest:
      type: object
      required:
        - word
      properties:
        word:
          type: string
          description: Слово или фраза, которую нужно исключить из топа
          minLength: 1
          example: casino
    StopWordResponse:
      type: object
      required:
        - word
      properties:
        word:
          type: string
          example: casino
    StopWordsResponse:
      type: object
      required:
        - words
      properties:
        words:
          type: array
          items:
            type: string
          example:
            - casino
            - spam
            - fake
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: invalid request